Add image tag to dc service definition
One issue is when trying to push the image in the bin/dev-env/build script:

Unable to fetch the defined image name in the docker-compose.yml without parsing the file using other tooling like yq.
So as a result, need to redefine the IMAGE_NAME in bin/dev-env/build.

Ideally can also use docker-compose push command - but perhaps can do this in the future